+++ email_archiver/__init__.py
@@ -1,2 +1,2 @@
 """Email-to-EML Secure Archiver"""
-__version__ = "1.2.8"
+__version__ = "1.3.0"